requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.0
filelock==3.18.0
frozenlist==1.4.1
fsspec==2025.3.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.1
langchain-community==0.3.1
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.129
MarkupSafe==3.0.2
marshmallow==3.22.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
openai==1.50.2
orjson==3.10.7
packaging==24.1
pillow==11.2.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.5.2
pydantic_core==2.27.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
yarl==1.13.1
